Move DB logic out of ManageRecordBadgesViewModel
